Don't assume stream_select will preserve array keys
It is not guaranteed that stream_select() will preserve the keys in the passed arrays. It seems that if the number of ready bytes is small, the keys are preserved, but for a larger numbers of bytes the array becomes 0-based.

With this patch, stream_select() is called with a copy of the $readSockets array and before returning the original keys are restored.

This fixes issue #30